+++ pydapper/postgresql/psycopg3.py
@@ -4,15 +4,12 @@ from pydapper.commands import Commands
 from pydapper.commands import CommandsAsync
 
 from .._context import CoroContextManager
-from ..main import register
-from ..main import register_async
 from ..utils import import_dbapi_module
 
 if TYPE_CHECKING:
     from ..dsn_parser import PydapperParseResult
 
 
-@register("psycopg")
 class Psycopg3Commands(Commands):
     @classmethod
     def connect(cls, parsed_dsn: "PydapperParseResult", **connect_kwargs) -> "Commands":
@@ -28,7 +25,6 @@ class Psycopg3Commands(Commands):
         return cls(conn)
 
 
-@register_async("psycopg")
 class Psycopg3CommandsAsync(CommandsAsync):
     @classmethod
     async def connect_async(cls, parsed_dsn: "PydapperParseResult", **connect_kwargs) -> "CommandsAsync":